Begins modification of heartbeat.
Instead of the heartbeat being just a ping, I want to use it to do
synchronization between replcias.
Each replica should send out its current vector clock, and as a
response, it should receive any transactions that the other replica has
that are newer than that timestamp.

At this point, the query for messages has been written, along with changing the
SubscribeToHeartBeat RPC to use bidirectional streaming.